requirements.txt
argostranslate==1.9.6
beautifulsoup4==4.13.3
datasets==3.5.0
ftfy==6.3.1
langdetect==1.0.9
libquadmath==1.26.4
matplotlib==3.10.0
pandas==2.2.3
pillow==11.1.0
scikit-learn==1.6.1
seaborn==0.13.2
spacy==3.8.4
torch==2.6.0+cu126
torchcam==0.4.0
torchvision==0.21.0
transformers==4.49.0
